test: isolate Standard Schema prototype security coverage
